backend/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "breast-cancer-detector"
version = "0.1.0"
description = "AI breast cancer risk classifier MVP"
authors = [{name = "Research Team"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "Django>=5.0",
    "djangorestframework>=3.14",
    "django-cors-headers>=4.0",
    "scikit-learn>=1.3",
    "numpy>=1.24",
    "pandas>=2.0",
    "python-dotenv>=1.0",
    "typing-extensions>=4.5",
    # "shap>=0.42",  # Uncomment when needed for advanced explanations
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-django>=4.5",
    "black>=23.0",
    "flake8>=6.0",
]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]

